application_code/scripts/requirements-browser-use.txt
browser-use==0.12.2
langchain-google-genai>=2.0.0
playwright>=1.58.0
python-dotenv>=1.0.1
google-genai>=1.65.0
pydantic>=2.4.2
